requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
pillow
tqdm
numpy
faiss-cpu
torch
open-clip-torch
google-cloud-storage
